implemented reverse_lazy instead of hardcoded url
if the url changes - we will have to change the success_url everywhere. Not good.

What is better is to use the "name" of the url, which will remain the same hopefully.

So now when the url changes - links don't break down.

DRY priciple, dont have to change everywhere, can do it in one place.