move requirements.txt to .dockerfile(also simplify it)
Tried doing context ".." and moving Dockerfile to the root directory, to then be able to fetch the Dockerfile with ../Dockerfile that lies in root directory - it works, but the Dockerfile lying in the root directory, which is NOT an actual Dockerfile that you would use to create a Docker container is a little bit confusing.

This Dockerfile is ONLY used for devcontainer, it does not COPY any files or anything like that.. so when a proper Docker container will be needed, another Dockerfile in the root directory will be created.